Move useDebouncedWindowSize hook and add useDebouncedMatchMobileQuery hook
