+++ src/app/content/highlights/components/DisplayNote.tsx
@@ -6,6 +6,7 @@ import styled, { css } from 'styled-components/macro';
 import Dropdown, { DropdownItem, DropdownList } from '../../../components/Dropdown';
 import Times from '../../../components/Times';
 import { textStyle } from '../../../components/Typography/base';
+import { useDebouncedWindowSize } from '../../../reactUtils';
 import theme from '../../../theme';
 import { mergeRefs } from '../../../utils';
 import { highlightStyles } from '../../constants';
@@ -13,7 +14,6 @@ import { query } from '../../search/selectors';
 import { tocOpen } from '../../selectors';
 import { focusHighlight } from '../actions';
 import { cardPadding, cardWidth } from '../constants';
-import { useDebouncedWindowSize } from './cardUtils';
 import Confirmation from './Confirmation';
 import MenuToggle, { MenuIcon } from './MenuToggle';
 import TruncatedText from './TruncatedText';
